WebScrape/requirements.txt
beautifulsoup4>=4.12.0
selenium>=4.15.0
requests>=2.31.0
lxml>=4.9.0
Pillow>=10.0.0
pyodbc>=5.0.0
python-dotenv>=1.0.0
tqdm>=4.66.0


